Добавил функции добавления и удаления в/из базы данных с правами доступа. Также добавил учет новых файлов и папок в базе данных с помощью этих функций. Теперь каждый новый загруженный файл или созданная папка будут появляться в базе данных прав доступа

+++ userFunctions.php
@@ -42,6 +42,7 @@ switch ($action) {
         break;
     case "changeDirectory":
         $action($dirName);
+        $path = $_SESSION['path'];
         break;
     default:
         exit();
@@ -50,22 +51,36 @@ switch ($action) {
 $_RESULT["window"] = newWindow($path);
 $_RESULT["space"] = $_SESSION['availablespace'];
 
-function createDirectory($path, $dirName)
-{
+function createDirectory($path, $dirName){
     // if (!checkAccessRights($path, $_SESSION['user'])){
     //     return false;
     // }
     $dir = $path . '/' . $dirName;
     if(!is_dir($dir)) {
-        mkdir($dir);
-        //$ini = parse_ini_file("database/mysql.ini");
-        //$mysql = mysqli_connect($ini['host'], $ini['user'], $ini['password'], $ini['database']);
+        $ini = parse_ini_file("database/mysql.ini");
+        $mysql = mysqli_connect($ini['host'], $ini['user'], $ini['password'], $ini['database']);
+        if (mkdir($dir)) {
+            if (addToAccessrights($mysql, $dir)) {
+                mysqli_close($mysql);
+                return true;
+            }
+            else{
+                mysqli_close($mysql);
+                rmdir($dir);
+                return false;
+            }
+        }
+        else{
+            mysqli_close($mysql);
+            return false;
+        }
     }
-    return true;
+    else{
+        return false;
+    }
 }
 
-function deleteDirectory($path, $dirName)
-{
+function deleteDirectory($path, $dirName){
     // if (!checkAccessRights($path, $_SESSION['user'])){
     //     return false;
     // }
@@ -74,13 +89,27 @@ function deleteDirectory($path, $dirName)
     if (!$mysql) {
         return false;
     }
-    removeDir($path . "/" . $dirName, $mysql);
-    mysqli_close($mysql);
-    return true;
+    else{
+        $dir = $path . "/" . $dirName;
+        if (removeFromAccessrights($mysql, $dir)) {
+            if (removeDir($dir, $mysql)) {
+                mysqli_close($mysql);
+                return true;
+            } 
+            else {
+                addToAccessrights($mysql, $dir);
+                mysqli_close($mysql);
+                return false;
+            }
+        }
+        else{
+            mysqli_close($mysql);
+            return false;
+        }
+    }
 }
 
-function uploadFile($path)
-{
+function uploadFile($path){
     // if (!checkAccessRights($path, $_SESSION['user'])){
     //     return false;
     // }
@@ -88,46 +117,59 @@ function uploadFile($path)
     if ($size > $_SESSION['availableSpace']){
         return false;
     }
-    $filePath = $path."/".$_FILES['file']['name'];
-    $fileName = preg_split("/\./uis", $_FILES['file']['name'])[0];
-    if (file_exists($filePath)){
-        $number = 1;
-        foreach(glob("$path/$fileName*") as $file){
-            preg_match("@"."$path/$fileName"."\((?<number>[0-9]*)\)@uis", $file, $arr);
-            if ($arr["number"] > $number){
-                $number = $arr["number"];
+    else{
+        $filePath = $path."/".$_FILES['file']['name'];
+        $fileName = preg_split("/\./uis", $_FILES['file']['name'])[0];
+        if (file_exists($filePath)){
+            $number = 1;
+            foreach(glob("$path/$fileName*") as $file){
+                preg_match("@"."$path/$fileName"."\((?<number>[0-9]*)\)@uis", $file, $arr);
+                if ($arr["number"] > $number){
+                    $number = $arr["number"];
+                }
+                elseif ($arr["number"] == $number){
+                    $number++;
+                }
             }
-            elseif ($arr["number"] == $number){
-                $number++;
-            }
+            $filePath = preg_replace("@"."$path/$fileName"."@uis", "$path/$fileName($number)", $filePath);
         }
-        $filePath = preg_replace("@"."$path/$fileName"."@uis", "$path/$fileName($number)", $filePath);
-    }
-
-    if (move_uploaded_file($_FILES['file']['tmp_name'], $filePath)) {
         $ini = parse_ini_file("database/mysql.ini");
         $mysql = mysqli_connect($ini['host'], $ini['user'], $ini['password'], $ini['database']);
-        if (!$mysql) {
+        if (move_uploaded_file($_FILES['file']['tmp_name'], $filePath)) {
+            if (addToAccessrights($mysql, $filePath)) {
+                if (newAvailableSpace($size, "+", $_SESSION['username'], $mysql)){
+                    mysqli_close($mysql);
+                    return true;
+                }
+                else{
+                    removeFromAccessrights($mysql, $filePath);
+                    mysqli_close($mysql);
+                    unlink($filePath);
+                    return false;
+                }
+            }
+            else{
+                mysqli_close($mysql);
+                unlink($filePath);
+                return false;
+            }
+        }
+        else{
+            mysqli_close($mysql);
             return false;
         }
-        newAvailableSpace($size, "+", $_SESSION['username'], $mysql);
-        mysqli_close($mysql);
     }
-
-    return true;
 }
 
 
-function downloadFile($path, $fileName)
-{
+function downloadFile($path, $fileName){
     // if (!checkAccessRights($path, $_SESSION['user'])){
     //     return false;
     // }
     return true;
 }
 
-function deleteFile($path, $fileName)
-{
+function deleteFile($path, $fileName){
     // if (!checkAccessRights($path, $_SESSION['user'])){
     //     return false;
     // }
@@ -138,19 +180,35 @@ function deleteFile($path, $fileName)
     if (!$mysql) {
         return false;
     }
-    if (newAvailableSpace($size, "-", $_SESSION['username'], $mysql)) {
-        mysqli_close($mysql);
-        unlink($file);
-        return true;
-    }
     else{
-        return false;
+        if (newAvailableSpace($size, "-", $_SESSION['username'], $mysql)) {
+            if (removeFromAccessrights($mysql, $file)){
+                if (unlink($file)){
+                    mysqli_close($mysql);
+                    return true;
+                }
+                else{
+                    addToAccessrights($mysql, $file);
+                    newAvailableSpace($size, "+", $_SESSION['username'], $mysql);
+                    mysqli_close($mysql);
+                    return false;
+                }
+            }
+            else{
+                newAvailableSpace($size, "+", $_SESSION['username'], $mysql);
+                mysqli_close($mysql);
+                return false;
+            }
+        }
+        else{
+            mysqli_close($mysql);
+            return false;
+        }
     }
 }
 
 
-function changeMod($path, $newMod)
-{
+function changeMod($path, $newMod){
     // if (!checkAccessRights($path, $_SESSION['user'])){
     //     return false;
     // }
@@ -158,8 +216,7 @@ function changeMod($path, $newMod)
     return true;
 }
 
-function changeDirectory($dirName)
-{
+function changeDirectory($dirName){
     // if (!checkAccessRights($path, $_SESSION['user'])){
     //     return false;
     // }
@@ -167,32 +224,62 @@ function changeDirectory($dirName)
     return true;
 }
 
-function goBack($path)
-{
+function goBack($path){
     if (preg_match_all("/\//uis", $path) == 1) {
         return false;
     }
-    preg_match("/(?<newPath>.*)\/.*?$/uis", $path, $arr);
-    $_SESSION['path'] = $arr['newPath'];
-    return true;
+    else{
+        preg_match("/(?<newPath>.*)\/.*?$/uis", $path, $arr);
+        $_SESSION['path'] = $arr['newPath'];
+        return true;
+    }
 }
 
-function removeDir($path, $mysql)
-{
+function removeDir($path, $mysql){
     foreach (glob($path . '/*') as $file) {
         if (is_dir($file)) {
-            removeDir($file, $mysql);
-            rmdir($file);
+            if (removeDir($file, $mysql)){
+                continue;
+            }
+            else{
+                return false;
+            }
         } 
         else {
             $size = filesize($file);
-            if (unlink($file)) {
-                newAvailableSpace($size, "-", $_SESSION['username'], $mysql);
+            if (newAvailableSpace($size, "-", $_SESSION['username'], $mysql)) {
+                if (removeFromAccessrights($mysql, $file)){
+                    if (unlink($file)) {
+                        continue;
+                    }
+                    else{
+                        addToAccessrights($mysql, $file);
+                        newAvailableSpace($size, "+", $_SESSION['username'], $mysql);
+                        return false;
+                    }
+                }
+                else{
+                    newAvailableSpace($size, "+", $_SESSION['username'], $mysql);
+                    return false;
+                }
+            }
+            else{
+                return false;
             }
         }
     }
-    rmdir($path);
-    return true;
+    if (removeFromAccessrights($mysql, $path)){
+        if (rmdir($path)){
+            return true;
+        }
+        else{
+            addToAccessrights($mysql, $path);
+            return false;
+        }
+    }
+    else{
+        return false;
+    }
 }
 
 function newAvailableSpace($size, $sign, $username, $mysql){
@@ -208,7 +295,13 @@ function newAvailableSpace($size, $sign, $username, $mysql){
         $update = mysqli_query($mysql, "UPDATE `users` SET `availablespace`='$availablespace' WHERE `username`='$username'");
         if ($update) {
             $_SESSION['availablespace'] = $availablespace;
+            return true;
+        }
+        else{
+            return false;
         }
     }
-    return true;
+    else{
+        return false;
+    }
 }
